[core]: moved tests to seperate folder cli
